Adding Codeception Based Unit Testing and Updating PHPUnit Existing To Use New Traits (Squashed)
