+++ tests/unit/Config/LoaderTest.php
@@ -3,8 +3,7 @@
 namespace Phalcon\Test\Config;
 
 use Phalcon\Config\Loader as ConfigLoader;
-use Codeception\TestCase\Test;
-use UnitTester;
+use Phalcon\Test\Codeception\UnitTestCase as Test;
 
 /**
  * \Phalcon\Test\Config\LoaderTest
@@ -25,26 +24,6 @@ use UnitTester;
  */
 class LoaderTest extends Test
 {
-    /**
-     * UnitTester Object
-     * @var UnitTester
-     */
-    protected $tester;
-
-    /**
-     * executed before each test
-     */
-    protected function _before()
-    {
-    }
-
-    /**
-     * executed after each test
-     */
-    protected function _after()
-    {
-    }
-
     public function testLoadPhpFileConfig()
     {
         $file = INCUBATOR_FIXTURES . 'Config/config.php';
