+++ tests/unit/Mvc/Model/MetaData/BaseTest.php
@@ -2,8 +2,7 @@
 
 namespace Phalcon\Test\Mvc\Model\MetaData;
 
-use Codeception\TestCase\Test;
-use UnitTester;
+use Phalcon\Test\Codeception\UnitTestCase as Test;
 use Mockery;
 
 /**
@@ -25,26 +24,6 @@ use Mockery;
  */
 class BaseTest extends Test
 {
-    /**
-     * UnitTester Object
-     * @var UnitTester
-     */
-    protected $tester;
-
-    /**
-     * executed before each test
-     */
-    protected function _before()
-    {
-    }
-
-    /**
-     * executed after each test
-     */
-    protected function _after()
-    {
-    }
-
     public function testBaseMetaDataAdapterImplementsMetaDataInterface()
     {
         $mock = Mockery::mock('Phalcon\Mvc\Model\MetaData\Base');
