+++ tests/unit/Validation/Validator/AlphaCompleteValidatorTest.php
@@ -2,10 +2,8 @@
 
 namespace Phalcon\Test\Validation\Validator;
 
-use UnitTester;
 use Phalcon\Validation;
-use Codeception\TestCase\Test;
-use Phalcon\Validation\Validator\CardNumber;
+use Phalcon\Test\Codeception\UnitTestCase as Test;
 
 /**
  * \Phalcon\Test\Validation\Validator\AlphaCompleteValidatorTest
@@ -26,11 +24,6 @@ use Phalcon\Validation\Validator\CardNumber;
  */
 class AlphaCompleteValidatorTest extends Test
 {
-    /**
-     * UnitTester Object
-     * @var UnitTester
-     */
-    protected $tester;
 
     public function testAlphaCompleteValidatorOk()
     {
