+++ tests/unit/Validation/Validator/AlphaNamesValidatorTest.php
@@ -2,10 +2,8 @@
 
 namespace Phalcon\Test\Validation\Validator;
 
-use UnitTester;
 use Phalcon\Validation;
-use Codeception\TestCase\Test;
-use Phalcon\Validation\Validator\CardNumber;
+use Phalcon\Test\Codeception\UnitTestCase as Test;
 
 /**
  * \Phalcon\Test\Validation\Validator\AlphaNamesValidatorTest
@@ -26,12 +24,6 @@ use Phalcon\Validation\Validator\CardNumber;
  */
 class AlphaNamesValidatorTest extends Test
 {
-    /**
-     * UnitTester Object
-     * @var UnitTester
-     */
-    protected $tester;
-
     public function testNamesValidatorOk()
     {
         $data['text'] = 'Richard Feynman';
